+++ R/functions-MSnSet.R
@@ -26,6 +26,11 @@ normalise_MSnSet <- function(object, method, ...) {
     e <- exprs(object)
     center <- apply(e, 2L, median, na.rm = TRUE)
     e <- sweep(e, 2L, center, check.margin = FALSE, ...)
+  } else if (method == "diff.median") {
+      e <- exprs(object)
+      med <- median(as.numeric(e), na.rm = TRUE)
+      cmeds <- apply(e, 2L, median, na.rm = TRUE)
+      e <- sweep(e, 2L, cmeds - med)
   } else {
     switch(method,
            max = div <- apply(exprs(object), 1L, max, na.rm = TRUE),