+++ R/methods-MSnSet.R
@@ -87,8 +87,8 @@ setMethod("show","MSnSet",
 setMethod("normalize", "MSnSet",
           function(object,
                    method = c("sum", "max", "center.mean",
-                              "center.median", "quantiles",
-                              "quantiles.robust", "vsn"), ...)
+                              "center.median", "diff.median",
+                              "quantiles", "quantiles.robust", "vsn"), ...)
               normalise_MSnSet(object, match.arg(method), ...)
           )