+++ NEWS.md
@@ -1,7 +1,7 @@
 # MSnbase 2.1
 
 ## Changes in version 2.1.10
-- (nothing yet)
+- New `diff.median` normalisation for `MSnSet`s. <2017-01-26 Thu>
 
 ## Changes in version 2.1.9
 - When fully trimmed, an (empty) spectrum has peaksCount of 0L - see